[ feature_01 ]
file upload 기초 셋팅

+++ src/main/java/com/fsje/dairy/dto/DiaryDto.java
@@ -11,7 +11,13 @@ import lombok.Data;
  */
 @Data
 public class DiaryDto {
-	private long diaryId;
+	/*
+	 * int, long -> Interger, Long 타입을 선언한 이유 : 
+	 * int, long은 primitive type이라 초기화시 0값을 가짐
+	 * 의도적으로 0값을 할당한건지 초기화된 0값인지 알 수 없으므로
+	 * wrapper class(Integer, Long)를 사용해 초기화시 null이 지정되게함
+	 */
+	private Integer diaryId;
 	private String diaryTitle;
 	private String diaryContent;
 	private String regUser;
